pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ruff>=0.6.2",
    "modal>=0.64.51",
    "fastapi>=0.112.1",
    "uvicorn>=0.30.6",
    "sqlalchemy>=2.0.32",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.9",
    "greenlet>=3.0.3",
    "python-jose>=3.3.0",
    "pytest-watch>=4.2.0",
    "pytest>=8.3.2",
    "httpx>=0.27.0",
    "protobuf>=4.25.4",
    "pytest-asyncio>=0.24.0",
    "trio>=0.26.2",
    "logfire[fastapi]>=0.51.0",
    "opentelemetry-instrumentation-sqlalchemy>=0.47b0",
    "opentelemetry-instrumentation-asyncio>=0.47b0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
